extracted helper
