+++ app/react/Settings/components/Dashboard.tsx
@@ -2,16 +2,9 @@ import React, { useEffect, useState } from 'react';
 import { connect, ConnectedProps } from 'react-redux';
 import { Translate } from 'app/I18N';
 import { IStore } from 'app/istore';
+import { formatBytes } from 'V2/shared/formatHelpers';
 import { SettingsHeader } from './SettingsHeader';
 import SettingsAPI from '../SettingsAPI';
-
-const formatBytes = (bytes: number) => {
-  //Sourced from https://stackoverflow.com/questions/15900485
-  if (!+bytes || +bytes < 0) return '0 Bytes';
-  const sizes = ['Bytes', 'KB', 'MB', 'GB', 'TB', 'PB', 'EB', 'ZB', 'YB'];
-  const index = Math.floor(Math.log(bytes) / Math.log(1024));
-  return `${parseFloat((bytes / 1024 ** index).toFixed(2))} ${sizes[index]}`;
-};
 
 const mapStateToProps = (state: IStore) => ({
   stats: state.settings.stats,
